- Moved projects to (modules)/projects.
