+++ src/app/admin/(modules)/projects/_components/Loading/ScreenshotItemLoading.tsx
