+++ src/app/admin/(modules)/projects/_components/ScreenshotDialog.tsx
